requirements.txt
rembg[cpu]
Pillow>=10.0.0
flask>=3.0.0
onnxruntime>=1.16.0
